+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 # develop
 
+Added:
+
+- Use encoding specified in meta tag of HTML document when server does not
+  specify encoding. [#52]
+
+[#52]: https://github.com/pdc/linotak/issues/52
+
 ## 0.3.1 (2025-10-19)
 
 Deletions: